+++ src/index.ts
@@ -1,1 +1,5 @@
-export const VERSION = "0.1.0";
+import "./global.css";
+
+export { VERSION } from "./version";
+export { ThemeProvider, useTheme } from "./theme/ThemeProvider";
+export type { ThemeChoice, ResolvedTheme, ThemeContextValue, ThemeProviderProps } from "./theme/ThemeProvider";